Create 5. FRIEND Class, DEFFER Keyword:
4. FRIEND Class, DEFFER Keyword:

In general, the protected and private components of a class cannot be accessed outside the class. However, in certain cases, there may be a business requirement to access these components. 
This can be achieved using the FRIEND class feature in Object-Oriented Programming (OOP), which allows specific classes to access all components of another class, regardless of their visibility section (protected or private).
The DEFFER keyword is used to inform the system that the class is declared later in the program, rather than at the beginning.