7. Create Abstract Class
7. Abstract Class in ABAP
A class that contains at least one abstract method is called an abstract class. An abstract method is a method that is declared but not implemented in the class. We declare a method as abstract when we do not know its implementation details yet. The actual implementation of the abstract method is provided in the subclass.
Abstract methods are declared using the ABSTRACT keyword. If a class contains at least one abstract method, the class itself must also be declared as abstract. Abstract methods should be declared in the public or protected section of the class. They are also referred to as non-concrete methods.
Static methods and constructors cannot be declared as abstract because they cannot be redefined in a subclass.
Any subclass that inherits from an abstract class must implement all of the abstract methods of the superclass. If a subclass does not implement all abstract methods, it too must be declared as abstract.
Abstract classes cannot be instantiated directly because they are incomplete. However, once the abstract methods are implemented in the subclass, we can create an object of the subclass.
Assigning a subclass object to a reference of the abstract superclass is known as narrow casting. For example, <Abs_spr_cls_ref> = <Obj_sub_cls>.
When narrow casting is done, the reference variable can access only the methods of the superclass. To access methods specific to the subclass, we need to use dynamic method calls.
Narrow casting refers to the process of assigning a more specific object (subclass) to a reference of a more general type (abstract superclass), essentially switching from a more detailed view to a less detailed one.