Create Interface
8.Interface
Pure Abstract Class:
An interface in ABAP is often referred to as a "pure abstract class." This means that all methods defined within an interface are abstract by default. Interfaces do not contain any method implementations, only method declarations.

Public Visibility:
Interface components, such as constants and methods, are public by default. An interface does not have an explicit visibility section like a normal class. Therefore, any method or constant declared in an interface is automatically public and accessible to classes that implement the interface.

No Method Implementations:
Interfaces cannot contain method implementations. They only contain method declarations. The responsibility for providing the method implementation lies with the class that implements the interface.

Implementing an Interface:
If a class wants to implement an interface (whether local or global), it must declare the interface in the public section of the class using the INTERFACES keyword.
The class must also implement all the methods declared in the interface.

If the class does not implement all the methods, the class should be declared as abstract.
Methods that are not implemented in an abstract class should be flagged as abstract method.
A subclass can then inherit from the abstract class and provide implementations for the abstract methods.
Implementation Class:
A class that implements an interface is called the implementation class. This class provides the concrete implementations of the all the methods declared in the interface.

Interfaces must in the Public Section:
When implementing an interface, the interface must be declared in the public section of the local class. This ensures that the methods of the interface are accessible from outside the class.

Multiple Inheritance:
Interfaces allow for multiple inheritance. A single class can implement multiple interfaces, which means it can inherit method declarations from multiple sources. This is one of the main advantages of using interfaces in object-oriented programming.

Prefixing Interface Methods:
When referring to a method of an interface outside the context of the interface, the method must be prefixed with the name of the interface, followed by a tilde (~).

<interface_name>~<method_name>

This ensures that you are calling the method of the correct interface.

No Constructors in Interfaces:
Interfaces cannot contain constructors. Constructors are specific to classes, and since interfaces do not contain any implementation, they do not define any constructor logic.

Interface Inheritance:


In ABAP, an interface can include other interfaces. When a class implements an interface that has included interfaces, the class must implement all the methods from the parent interface as well as any methods from the included interfaces.
This means that if an interface inherits methods from other interfaces (whether directly or indirectly), the class implementing the main interface is required to implement all methods from all inherited interfaces, not just those defined in the immediate interface.
Global Classes Interfaces:
If a global class needs to implement one or more global interfaces, it must declare those interfaces in the INTERFACES tab.